+++ src/test/resources/jsonExamples/payloads/bulk/scan/basicForm.json
@@ -4,6 +4,6 @@
   "po_box": "PO 17",
   "ocr_data_fields": [
     { "name": "firstName", "value": "Christopher" },
-    { "name": "lastName", "value": "O'Jòhn" }
+    { "name": "lastName", "value": "O'John" }
   ]
 }
